Use specimen matching DB for specimen migration
When migrating specimen matches from ETS to NTBS, use the list of ETS
specimen matches stored in the specimen matching DB, rather than the
one stored in the migration DB, which is being removed as part of this
ticket NTBS-2279.
